+++ ui/hooks/index.ts
@@ -1,6 +1,6 @@
 import { isAllowedQueryParamPage } from "@tallyho/provider-bridge-shared"
 
-import { useState, useEffect, useRef, ReactElement, ReactNode } from "react"
+import { useState, useEffect, ReactElement, ReactNode } from "react"
 import SharedPanelSwitcher from "../components/Shared/SharedPanelSwitcher"
 
 export * from "./redux-hooks"
@@ -23,29 +23,6 @@ export function useIsDappPopup(): boolean {
   }, [])
 
   return isDappPopup
-}
-
-export function useRunOnFirstRender(func: () => void): void {
-  const isFirst = useRef(true)
-
-  if (isFirst.current) {
-    isFirst.current = false
-    func()
-  }
-}
-
-export function useSkipFirstRenderEffect(
-  func: () => void,
-  deps: unknown[] = []
-): void {
-  const didMount = useRef(false)
-
-  useEffect(() => {
-    if (didMount.current) func()
-    else didMount.current = true
-    // We are passing in the dependencies when we initialize this hook, so we can not know what it will be exactly and it's ok.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, deps)
 }
 
 /**
