+++ ui/hooks/react-hooks.ts
@@ -83,3 +83,26 @@ export function useSetState<S extends Record<string, unknown>>(
 
   return [value, setter] as const
 }
+
+export function useRunOnFirstRender(func: () => void): void {
+  const isFirst = useRef(true)
+
+  if (isFirst.current) {
+    isFirst.current = false
+    func()
+  }
+}
+
+export function useSkipFirstRenderEffect(
+  func: () => void,
+  deps: unknown[] = []
+): void {
+  const didMount = useRef(false)
+
+  useEffect(() => {
+    if (didMount.current) func()
+    else didMount.current = true
+    // We are passing in the dependencies when we initialize this hook, so we can not know what it will be exactly and it's ok.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, deps)
+}
